Creación de roles y permisos
No pueden ver reportes , bancos , reportes , ni editar , ni eliminar

+++ database/migrations/tenant/2019_02_20_164002_tenant_pos_table.php
@@ -18,7 +18,8 @@ class TenantPosTable extends Migration
             $table->increments('id');
 
             // unity
-            $table->integer('user_id')->unsigned();
+            // $table->integer('user_id')->unsigned();
+            $table->bigInteger('user_id')->unsigned();
             $table->integer('establishment_id')->unsigned();
 //            $table->char('name', 50);
